Add account page for username actions

+++ backend/controllers/users.js
@@ -123,10 +123,13 @@ exports.changePassword = async (req, res) => {
 exports.changePhoto = async (req, res) => {
     let user = await db.getUserById(req.user.userId);
     let fileName = user.photo.split("/");
-    fileName = fileName[fileName.length - 1];
-    let pathImage = path.normalize(__dirname + `/../images/avatars/${fileName}`);
-    fs.rmSync(pathImage);
-    await db.changePhoto(req.user.userId, `http://localhost:3000/public/avatars/${req.file.filename}`);
+    if(user.photo != `http://${process.env.DOMAIN}/public/avatars/none.webp` && req.file.fileName == fileName){
+        let fileName = user.photo.split("/");
+        fileName = fileName[fileName.length - 1];
+        let pathImage = path.normalize(__dirname + `/../images/avatars/${fileName}`);
+        fs.rmSync(pathImage);
+    }
+    await db.changePhoto(req.user.userId, `http://${process.env.DOMAIN}/public/avatars/${req.file.filename}`);
     res.send({message: "Photo uploaded"});
 }
